Use props in the Product component
Until now we were using static text in the Product component, this
won't be very useful for a real world store front with many
products, where prices and descriptions change constantly. In a real
application there would be an admin portal where the owner of the
store could add products to his catalog, upload images, edit
descriptions, edit titles, and so on.

What we really need to do, is collect all the information the
store clerk enters in their admin portal, save it somewhere,
ideally in a database where we can easily access it, update it,
and finally, show it to the world on our store front site.

What we need to do now, is remove all the static text from
this component, instead we will use "props", which is passed
from one component to another, all the product data will be
loaded somehow (we'll get to it soon) and then passed in as
a single argument we will call "props", we will then replace
all the static text with the dynamic props fields.