Pass actual params
Now that we have prepared the Product component to accept product information, we can now go and pass in the props where we create the product component, in the Products component, the one where we list the products.

Now that we are using dynamic props (the text isn't hardcoded in the Product component but passed in from the Products list component), we can use different products for each line, we will get the same beautiful style and buttons, with a different image and text for each line.