Fetch from the backend
Now that we know how to use props with hardcoded data, let's move on to the next phase, fetching data from the backend and rendering it.

To do that we will need to use a "useState" to define our products, this is how we tell react to listen to every time we update the "products" variable using the "setProducts" function and rerender the page with the new data.

Here's what will happen, the App component will render the Products component, an empty array called products is created along with a function to update the products called "setProducts", we then use a "useEffect" which is triggered automatically when the component is rendered, in it we fetch the data, react will first render the page empty, but once we call setProducts and pass in our array of products that we got from the backend, react will update the page with the new list of Products.js

We can also use effects to listen to other events, such as user clicks and more, this is how we keep our components alive, ready for new data, ready to be played with.